requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.10
pydantic>=2.0.0
cdp-sdk>=0.1.0
python-dotenv>=1.0.0
gradio>=4.19.0 